Various improvements around HASH nodes
The example previously would lead to corrupt memory
when running for items that needed to re-hash the
hash array. This has been fixed.

Secondly, all HASH node memory leaks are now fixed,
that resulted from badly tracking objects through the
resize process. A new function `hash_free_shallow` was
added to help with this. The function `array_free_shallow`
is unused but might become useful in the future and so
was not removed.

The hash strategy is still garbage but this can be fixed later :)